Create SalesInvoiceAmount
Requesting all the Sales Invoice Lines from Business Central and printing the sales amount for each customer. 

This is a basic template that can be used for almost any table in the database, you'll just have to swap the field names and variables to get exactly what you are looking for.